# docs/openapi.yaml
# Здесь будет OpenAPI описание всех доступных эндпоинтов и моделей данных.
openapi: 3.0.3
info:
  title: Bank Cards API
  version: 1.0.0
  description: API для управления банковскими картами с поддержкой ролевой авторизации

servers:
  - url: http://localhost:8080
    description: Локальный сервер

tags:
  - name: Аутентификация
    description: Получение JWT-токена
  - name: Карты (ADMIN)
    description: Управление картами (только ADMIN)
  - name: Карты (USER)
    description: Просмотр и управление своими картами (USER)
  - name: Переводы
    description: Переводы между картами пользователя

paths:
  /auth/login:
    post:
      tags: [Аутентификация]
      summary: Аутентификация и получение JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Неверные учётные данные

  /api/admin/cards:
    post:
      tags: [Карты (ADMIN)]
      summary: Создать новую карту (ADMIN)
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardCreateRequest'
      responses:
        '201':
          description: Карта создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '403':
          description: Доступ запрещён (не ADMIN)

    get:
      tags: [Карты (ADMIN)]
      summary: Получить все карты (ADMIN)
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
      responses:
        '200':
          description: Список всех карт
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardPageResponse'

  /api/admin/cards/{id}/status:
    patch:
      tags: [Карты (ADMIN)]
      summary: Изменить статус карты (активировать/блокировать) (ADMIN)
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [ACTIVE, BLOCKED]
              required: [status]
      responses:
        '200':
          description: Статус обновлён
        '404':
          description: Карта не найдена

  /api/admin/cards/{id}:
    delete:
      tags: [Карты (ADMIN)]
      summary: Удалить карту (ADMIN)
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Карта удалена
        '404':
          description: Карта не найдена

  /api/admin/users:
    post:
      tags: [Карты (ADMIN)]
      summary: Создать пользователя (ADMIN)
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: Пользователь создан

  /api/cards:
    get:
      tags: [Карты (USER)]
      summary: Получить свои карты (USER)
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - name: maskedCardNumber
          in: query
          required: false
          schema:
            type: string
          description: Фильтр по маске карты (например, "****1234")
      responses:
        '200':
          description: Список карт пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardPageResponse'

  /api/cards/{id}:
    get:
      tags: [Карты (USER)]
      summary: Получить карту по ID (только свою)
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Карта найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '403':
          description: Нет доступа к чужой карте
        '404':
          description: Карта не найдена

    patch:
      tags: [Карты (USER)]
      summary: Запросить блокировку своей карты (USER)
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [BLOCKED]
              required: [status]
      responses:
        '200':
          description: Запрос на блокировку принят
        '403':
          description: Нет доступа
        '409':
          description: Карта уже заблокирована или неактивна

  /api/transfers:
    post:
      tags: [Переводы]
      summary: Перевод между своими картами (USER)
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '200':
          description: Перевод выполнен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResponse'
        '400':
          description: Недостаточно средств, неверные данные
        '403':
          description: Нет доступа к одной из карт

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
    size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100

  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password

    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT токен
        roles:
          type: array
          items:
            type: string
            enum: [ADMIN, USER]

    UserCreateRequest:
      type: object
      required: [username, password, email]
      properties:
        username:
          type: string
        password:
          type: string
        email:
          type: string
          format: email

    CardCreateRequest:
      type: object
      required: [cardholderName, maskedCardNumber, encryptedCardNumber, expiryMonth, expiryYear]
      properties:
        userId:
          type: integer
          format: int64
          description: ID владельца (для ADMIN)
        cardholderName:
          type: string
        maskedCardNumber:
          type: string
          example: "**** **** **** 1234"
        encryptedCardNumber:
          type: string
        expiryMonth:
          type: integer
          minimum: 1
          maximum: 12
        expiryYear:
          type: integer
          minimum: 2024
        balance:
          type: number
          format: decimal
          example: 1000.00

    CardResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        maskedCardNumber:
          type: string
        cardholderName:
          type: string
        expiryMonth:
          type: integer
        expiryYear:
          type: integer
        status:
          type: string
          enum: [ACTIVE, BLOCKED, EXPIRED]
        balance:
          type: number
          format: decimal
        createdAt:
          type: string
          format: date-time

    CardPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CardResponse'
        pageable:
          type: object
        totalPages:
          type: integer
        totalElements:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        size:
          type: integer
        number:
          type: integer

    TransferRequest:
      type: object
      required: [fromCardId, toCardId, amount]
      properties:
        fromCardId:
          type: integer
          format: int64
        toCardId:
          type: integer
          format: int64
        amount:
          type: number
          format: decimal
          minimum: 0.01

    TransferResponse:
      type: object
      properties:
        fromCardId:
          type: integer
          format: int64
        toCardId:
          type: integer
          format: int64
        amount:
          type: number
          format: decimal
        timestamp:
          type: string
          format: date-time